# be_agent_orchestrator/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Orchestrator",
    "description": "\nYuno AI Agent Orchestration Platform \u2014 backend API.\n\nStack: FastAPI \u00b7 LangGraph \u00b7 PostgreSQL \u00b7 Redis \u00b7 Arq.\n\n**Auth.** All non-public endpoints require an `Authorization: Bearer <access_token>`\nheader. Obtain tokens via `POST /api/v1/auth/login` and rotate via\n`POST /api/v1/auth/refresh`. The access token is valid for 12 hours; the\nrefresh token for 30 days.\n\n**Error envelope.** Every non-2xx response is `{\"error\": {\"code\", \"message\", \"details\"}}`.\nSwitch on `code` in the frontend (it's stable); render `message` to the user;\ninspect `details` for structured per-error data.\n",
    "version": "0.1.0",
    "contact": {
      "name": "Yuno AI",
      "email": "engineering@yuno.ai"
    },
    "license": {
      "name": "Proprietary"
    },
    "x-environment": "local"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    },
    {
      "url": "http://api:3001",
      "description": "Docker compose internal"
    }
  ],
  "paths": {
...
    "/api/v1/llm-providers/": {
      "get": {
        "tags": [
          "llm-providers"
        ],
        "summary": "List my LLM provider credential statuses",
        "description": "Returns one entry per supported provider (`openai`, `gemini`, `groq`), indicating whether the current user has saved credentials. The actual API keys are never returned.",
        "operationId": "llm_providers_get_list",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMCredentialsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/llm-providers/{provider}/credentials": {
      "get": {
        "tags": [
          "llm-providers"
        ],
        "summary": "Get credential status for one provider",
        "description": "Returns the configured-status + non-secret metadata for one provider. If unconfigured, `is_configured` is `false` and other fields are null.",
        "operationId": "llm_providers_get_credentials_status",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Provider key. One of `openai`, `gemini`, `groq`.",
              "title": "Provider"
            },
            "description": "Provider key. One of `openai`, `gemini`, `groq`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMCredentialStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "llm-providers"
        ],
        "summary": "Set / update LLM provider credentials",
        "description": "Upsert credentials for one LLM provider. Returns 204 with no body \u2014 the frontend's `llmProviderApi.registerCredential` parses with `z.void()`. Refetch via `GET /llm-providers/{provider}/credentials` if you need the updated status. The submitted payload fully replaces existing credentials (no deep merge); the actual key is stored in Postgres JSONB, never echoed back, and redacted in HTTP access logs.",
        "operationId": "llm_providers_post_credentials",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Provider key.",
              "title": "Provider"
            },
            "description": "Provider key."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLLMCredentialsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "llm-providers"
        ],
        "summary": "Remove LLM provider credentials",
        "description": "Deletes the user's stored credentials for one provider. Idempotent \u2014 deleting non-existent credentials returns 204.",
        "operationId": "llm_providers_delete_credentials",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Provider key.",
              "title": "Provider"
            },
            "description": "Provider key."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable upper-snake error code. Switch on this in the UI."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable summary; safe to surface to the user."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Structured per-error data (validation errors, conflict keys, etc)."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "The `error` field of every non-2xx response.\n\nMirrors what ``app.core.errors.register_exception_handlers`` emits, so\nthe OpenAPI schema matches the real wire format byte-for-byte.",
        "examples": [
          {
            "code": "AGENT_NOT_FOUND",
            "details": {
              "agent_id": "01HZ..."
            },
            "message": "Agent not found"
          }
        ]
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Top-level error response shape returned by every error path.",
        "examples": [
          {
            "error": {
              "code": "VALIDATION_ERROR",
              "details": {
                "errors": [
                  {
                    "loc": [
                      "body",
                      "email"
                    ],
                    "msg": "field required"
                  }
                ]
              },
              "message": "Invalid request payload"
            }
          }
        ]
      },
...
      "LLMCredentialStatus": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "is_configured": {
            "type": "boolean",
            "title": "Is Configured"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "When the credential was last set/updated. Null if unconfigured."
          }
        },
        "type": "object",
        "required": [
          "provider",
          "is_configured"
        ],
        "title": "LLMCredentialStatus",
        "description": "Status of one provider's credential for the current user.\n\nNever echoes the `api_key`. Surfaces only the public fields the FE\nneeds to render the settings card: provider, configured flag, the\nnon-secret bits (base_url, organization), and timestamps.",
        "examples": [
          {
            "is_configured": true,
            "provider": "openai",
            "updated_at": "2026-05-24T08:15:30.000Z"
          },
          {
            "is_configured": false,
            "provider": "gemini"
          }
        ]
      },
      "LLMCredentialsListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LLMCredentialStatus"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LLMCredentialsListResponse",
        "description": "List of credential statuses, one entry per supported provider."
      },
...
      "SetLLMCredentialsRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key",
            "description": "Provider API key (e.g. `sk-...`, `gsk_...`, `AIza...`)."
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Custom API base URL. Useful for OpenAI-compatible endpoints / Groq mirrors."
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization",
            "description": "OpenAI organization id, if applicable."
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "SetLLMCredentialsRequest",
        "description": "Set or update credentials for a single LLM provider.\n\nUpsert semantics: if the user already has a row for this provider,\nthe existing row's `credentials` JSONB is overwritten with the new\npayload (use full replacement, not deep-merge \u2014 clearer behavior).",
        "examples": [
          {
            "api_key": "sk-proj-..."
          },
          {
            "api_key": "gsk_..."
          },
          {
            "api_key": "AIza..."
          }
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste your access token (without the `Bearer ` prefix)."
      }
    }
  },
...
}